requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.2.1
MarkupSafe==2.1.5
python-decouple==3.8
setuptools==68.2.2
Werkzeug==3.0.2
wheel==0.41.2
